Hack type constructors for fileReader/Writer
Modifies the type constructor implementation to allow for a fake 'kind'
field in fileReader and fileWriter. If a named 'kind' expression exists,
it will be renamed to point to the temporary '_kind' field. If a named
expression does not exist, then the first non-named actual is bound to
the '_kind' field.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>