Dyno: Relax unknown-formal rule to enable more initial signature patterns
Closes https://github.com/Cray/chapel-private/issues/7398.

Chapel's function resolution proceeds in two phases:

1. A simple "zippering" of formals and actuals, in which
   each pair is independently checked for compatibility
2. A dependency-aware sequential traversal in which type
   information (e.g., type queries) from all preceding
   formal/actual pairs is carried forward into subsequent
   ones.

For the initial phase, it's entirely possible for some formal
type expressions to be missing information (e.g., if we have
`formal2: f(formal1.type)`. In this case, the existing call
resolution logic marks `f(...)` as having unknown type, and
allows actuals to be passed to it. More strict checking is
deferred until phase 2. However, prior to this PR, this
allowing-unknown-formals was not done for `type` formals. It's
unclear to me why that is the case. As a result, formals like
`type formal2: f(formal1.type)` did not work. This PR relaxes
the `canPass` condition to allow passing to `unknown` type formals
as well. This doesn't break any tests, but enables generic cast
operators:

```Chapel
operator :(x, type t: Wrapper(x.type)) {
  var w: Wrapper(x.type) = x;
  return w;
}
```

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>